Blade components, form request validation, forgot password integrated

+++ app/Http/Controllers/StudentController.php
@@ -2,53 +2,30 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StudentStoreValidation;
+use App\Http\Requests\StudentUpdateValidation;
 use App\Models\Student;
 use Illuminate\Http\Request;
 use Brian2694\Toastr\Facades\Toastr;
 
 class StudentController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         $students = Student::all();
         return view('admin.student.student-index', compact('students'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         return view('admin.student.create-student');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-
-    public function store(Request $request)
+    public function store(StudentStoreValidation $request)
     {
-        $request->validate([
-            'name' => 'required|string|min:3|max:35',
-            'email' => 'required|email|max:255|unique:students,email',
-            'avatar' =>  'required|image|mimes:jpeg,png,jpg,gif',
-            'password'  => 'required|min:6|max:32',
-            'password_confirmation' => 'required|min:6|same:password'
 
-        ]);
-
-        if ($request->hasFile('avatar')){
-            $avatar = $request->avatar->move('backend/admin/images/', $request->avatar->hashName());
+        if ($request->hasFile('avatar')) {
+            $avatar = $request->avatar->move('backend/uploads/', $request->avatar->hashName());
         }
 
 
@@ -59,29 +36,12 @@ class StudentController extends Controller
             'password' => $request->password
         ]);
 
-        if($student){
+        if ($student) {
             Toastr::success('WOW! New Student added successfully!', 'Created', ["positionClass" => "toast-top-center"]);
         }
         return redirect()->route('students.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Student  $student
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Student $student)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Student  $student
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Student $student)
     {
         $studentData = $student;
@@ -95,61 +55,38 @@ class StudentController extends Controller
      * @param  \App\Models\Student  $student
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Student $student)
+    public function update(StudentUpdateValidation $request, Student $student)
     {
-        // Data Validation
-        $request->validate([
-            'name' => 'required|string|min:2|max:35',
-            'email' => 'required|email|max:255|unique:students,email,'.$student->id,
-        ]);
+        if ($request) {
+            // updating data
+            $student->update([
+                'name' => $request->name,
+                'email' => $request->email,
+            ]);
+        }
 
         // Validate password if want to change
-        if ($request->change_password == 1) {
-            $this->validate($request, [
-                'password'  => 'required|min:6',
-                'password_confirmation' => 'required|min:6|same:password'
+        if ($request->change_password === 1) {
+            // Updating password
+            $student->update([
+                'password' => $request->password
             ]);
         }
 
         // Validate image if want to profile image
         if ($request->hasFile('avatar')) {
-            $request->validate([
-                'avatar' =>  'image|mimes:jpeg,png,jpg,gif'
-            ]);
-        }
 
-        if ($request->hasFile('avatar')) {
             $oldPicture = $student->avater;
-            if (file_exists($oldPicture)) {
-                if ($oldPicture != 'backend/images/default.png') {
-                    unlink($oldPicture);
-                }
+            if (file_exists($oldPicture) && $oldPicture != 'backend/admin/images/default.png') {
+                // unlink old image
+                unlink($oldPicture);
             }
-            $avatar = $request->avatar->move('backend/admin/images/', $request->avatar->hashName());
-        }
-        if ($request->hasFile('avatar') && $request->filled('password')){
+            // save file to host
+            $avatar = $request->avatar->move('backend/uploads/', $request->avatar->hashName());
+
+            //change avatar
             $student->update([
-                'name' => $request->name,
-                'email' => $request->email,
-                'avatar' => $avatar,
-                'password' => $request->password
-            ]);
-        } elseif($request->hasFile('avatar')){
-            $student->update([
-                'name' => $request->name,
-                'email' => $request->email,
                 'avatar' => $avatar
-            ]);
-        } elseif ($request->filled('password')) {
-            $student->update([
-                'name' => $request->name,
-                'email' => $request->email,
-                'password' => $request->password
-            ]);
-        } else {
-            $student->update([
-                'name' => $request->name,
-                'email' => $request->email,
             ]);
         }
 
@@ -157,21 +94,14 @@ class StudentController extends Controller
         return redirect()->route('students.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Student  $student
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(Student $student)
     {
-        // going to delete student data
-        if(file_exists('backend/admin/images/'.$student->image) AND !empty($student->image)){
-            unlink('backend/admin/images/'.$student->image);
+        if ($student->avatar && $student->avatar != 'backend/admin/images/default.png') {
+            unlink($student->avatar);
         }
-        $student->delete();
-        Toastr::warning('Student deleted successfully!', 'Deleted!', ["positionClass" => "toast-top-center"]);
-        return redirect()->route('students.index');
 
+        $student->delete();
+        Toastr::warning('Student deleted successfully!', 'Deleted!', ["positionClass" =>"toast-top-center"]);
+        return redirect()->route('students.index');
     }
 }
